Fix Linux build blocker, ID-truncation lookup bug, and double-.json save bug found in scheduled code scan #15
- WorldEditor.h included <Helpers.h> while the file on disk is
  helpers.h; case-sensitive filesystems (Linux/macOS) failed to
  compile. Confirmed by a clean CMake configure/build in this
  environment before the fix, and a full compile pass afterward.
- FindGameObjectByID/FindEntityByID/FindInteractableByID (in both
  GameMap and the WorldEditor free-function duplicates) took the id
  as uint8_t, truncating the uint64_t object/entity/interactable ids.
  Any id above 255 would silently fail to look up or alias onto the
  wrong object — this fed straight into WorldEditor::getSelectedObject
  and the interactable activator lookup. Restored uint64_t.
- SaveGame/LoadGame append ".json" to the filename internally, but
  mainMenu.cpp and gameMain.cpp passed "mainMenu.json"/"backup.json",
  producing "mainMenu.json.json" / "backup.json.json" — the exact
  issue flagged (but left unresolved) in the "Find World Files"
  commit. Main menu's scene load and the on-close backup save were
  both silently failing/writing to the wrong file. Passed the bare
  name instead.

+++ src/gameLayer/WorldEditor.cpp
@@ -52,7 +52,7 @@ void WorldEditor::ShowEditorHub()
 	ImGui::End();
 }
 
-GameObject* FindGameObjectByID(uint8_t id)
+GameObject* FindGameObjectByID(uint64_t id)
 {
 	const auto scene = SceneManager::getInstance().currentScene;
 	for (size_t i = 0; i < scene->gameMap.gameObjects.size(); ++i) {
@@ -62,7 +62,7 @@ GameObject* FindGameObjectByID(uint8_t id)
 	return nullptr;
 };
 
-Entity* FindEntityByID(uint8_t id)
+Entity* FindEntityByID(uint64_t id)
 {
 	const auto scene = SceneManager::getInstance().currentScene;
 	auto it = scene->entities.find(id);
@@ -70,7 +70,7 @@ Entity* FindEntityByID(uint8_t id)
 	return it->second.get();
 };
 
-InteractableObject* FindInteractableByID(uint8_t id)
+InteractableObject* FindInteractableByID(uint64_t id)
 {
 	const auto scene = SceneManager::getInstance().currentScene;
 	auto it = scene->interactables.find(id);
